py-functions/requirements.txt
firebase_functions
firebase-admin
google-cloud-logging